✨ feat(Discourse): Add DataComponent to display users and topics
The `DataComponent` is added to the `Discourse` component to display the list of users and topics. This component receives the `data` prop containing the user and topic data.

Before:
- No `DataComponent` in the `Discourse` component.

After:
- Added the `DataComponent` to the `Discourse` component.
- Imported the necessary components and types.

## src/components/Discourse/DataComponent.tsx
✨ feat(DataComponent): Create DataComponent to display users and topics

This commit adds the `DataComponent` to the `Discourse` folder. The `DataComponent` is responsible for rendering the list of users and topics. It receives the `data` prop of type `Data` which contains the necessary user and topic data.